Create README.md
my background of 15 years in horticulture. However, this is an industry I have been led to take on, due to life circumstances. I have in the last few months, now got the opportunity to follow a path that I wished to follow when I left high school and arrived in Australia in 2007.